Add flag for alternative pl/pgsql stored function directory
The pgsql-load-stored-procs script by default only looked in the share
directory location specified by pg_config.  This allows the scripts to
load pl/pgsql stored functions from a specified directory in the event
that the stored functions are not installed into the postgres share
directory.

Also add a flag to pgsql-build-db so that it can passed onto
pgsql-load-stored-procs.