Merge easy script behavior into run script
Make the "easy" script behavior the default behavior in the "run"
script.  Refactor the option handling so that long parameter names can
be used.  Also add a man page for the "run" script.

Additionally set more defaults to make the test easier to try with
having to set fewer options:

* default database name is dbt2
* default number of warehouses is 1
* default test duration is 3 minutes (180), so there will be at least 3
  data points collected
* default number of database connections is 10, for driver-mode 1